hugging_face_app/requirements.txt
fastapi
uvicorn
torch
open_clip_torch
pillow
requests
python-multipart
transformers
ftfy